cocomo jetzt richtig verbessert

+++ src/model/services/calculation/Cocomo.java
@@ -41,16 +41,16 @@ public class Cocomo extends Calculation {
 		switch (method) {
 		
 		case Embedded:
-			result.setPersonMonths(3.6 * Math.pow(linesOfCode/1000,1.2));
-			result.setTimeToDev(2.5 * Math.pow(linesOfCode/1000, 0.32));
+			result.setPersonMonths(3.6 * Math.pow((float)linesOfCode/1000,1.2));
+			result.setTimeToDev(2.5 * Math.pow((float)linesOfCode/1000, 0.32));
 			break;
 		case Organic:
-			result.setPersonMonths(2.4 * Math.pow(linesOfCode/1000,1.05));
-			result.setTimeToDev(2.5 * Math.pow(linesOfCode/1000, 0.38));
+			result.setPersonMonths(2.4 * Math.pow((float)linesOfCode/1000,1.05));
+			result.setTimeToDev(2.5 * Math.pow((float)linesOfCode/1000, 0.38));
 			break;
 		case Semidetached:
-			result.setPersonMonths(3.0 * Math.pow(linesOfCode/1000,1.12));
-			result.setTimeToDev(2.5 * Math.pow(linesOfCode/1000, 0.35));
+			result.setPersonMonths(3.0 * Math.pow((float)linesOfCode/1000,1.12));
+			result.setTimeToDev(2.5 * Math.pow((float)linesOfCode/1000, 0.35));
 			break;	
 		default:
 			result.setPersonMonths(0);
